requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
python-dotenv==1.0.1
openai==1.51.0
httpx==0.27.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
